techeducation_rag/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
autopep8==2.3.2
black==26.1.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.13.2
Flask==3.1.2
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.0
langchain-core==1.2.5
langchain-openai==1.1.6
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.2
MarkupSafe==3.0.3
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
numpy==2.4.2
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pathspec==1.0.4
platformdirs==4.5.1
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.1
pytokens==0.4.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.46
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
Werkzeug==3.1.5
xxhash==3.6.0
zstandard==0.25.0
